refactor: CaffeineCacheBuilder 최상위 함수를 사용한 빌더로 리팩터링
